models/surya-ocr-api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pillow>=10.0.0
pydantic>=2.0.0

# Surya OCR
surya-ocr>=0.6.0
torch>=2.0.0
torchvision>=0.15.0
